allow using as an instance
